dev: rename services/SurveyUpdater to services/SurveyAggregateService
