+++ application/models/services/SurveyAggregateService/LanguageConsistency.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace LimeSurvey\Models\Services\SurveyUpdater;
+namespace LimeSurvey\Models\Services\SurveyAggregateService;
 
 use Survey;
 use LSYii_Application;
