+++ tests/unit/api/opHandlers/OpHandlerSurveyUpdateTest.php
@@ -8,7 +8,7 @@ use LimeSurvey\ObjectPatch\Op\OpInterface;
 use LimeSurvey\ObjectPatch\Op\OpStandard;
 use LimeSurvey\ObjectPatch\OpHandler\OpHandlerException;
 use ls\tests\TestBaseClass;
-use ls\tests\unit\services\SurveyUpdater\GeneralSettings\GeneralSettingsMockSetFactory;
+use ls\tests\unit\services\SurveyAggregateService\GeneralSettings\GeneralSettingsMockSetFactory;
 
 class OpHandlerSurveyUpdateTest extends TestBaseClass
 {
