+++ tests/unit/services/SurveyAggregateService/GeneralSettings/GeneralSettingsFactory.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace ls\tests\unit\services\SurveyUpdater\GeneralSettings;
+namespace ls\tests\unit\services\SurveyAggregateService\GeneralSettings;
 
-use LimeSurvey\Models\Services\SurveyUpdater\GeneralSettings;
+use LimeSurvey\Models\Services\SurveyAggregateService\GeneralSettings;
 
 /**
  * General Settings Mock Factory
