+++ tests/unit/services/SurveyAggregateService/GeneralSettings/GeneralSettingsMockSet.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace ls\tests\unit\services\SurveyUpdater\GeneralSettings;
+namespace ls\tests\unit\services\SurveyAggregateService\GeneralSettings;
 
 use Survey;
 use Permission;
 use LSYii_Application;
 use CHttpSession;
 use LimeSurvey\PluginManager\PluginManager;
-use LimeSurvey\Models\Services\SurveyUpdater\LanguageConsistency;
+use LimeSurvey\Models\Services\SurveyAggregateService\LanguageConsistency;
 use User;
 
 class GeneralSettingsMockSet
