+++ tests/unit/services/SurveyAggregateService/GeneralSettings/GeneralSettingsUpdateCaptchaTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ls\tests\unit\services\SurveyUpdater\GeneralSettings;
+namespace ls\tests\unit\services\SurveyAggregateService\GeneralSettings;
 
 use ls\tests\TestBaseClass;
